ssize_t was used incorrectly (only [-1, INT_MAX]

+++ src/batch_request.cpp
@@ -46,14 +46,14 @@ CassError cass_batch_add_statement(CassBatch* batch, CassStatement* statement) {
 
 namespace cass {
 
-ssize_t BatchRequest::encode(int version, BufferVec* bufs) const {
+int BatchRequest::encode(int version, BufferVec* bufs) const {
   if(version != 2) {
     return ENCODE_ERROR_UNSUPPORTED_PROTOCOL;
   }
   return encode_v2(bufs);
 }
 
-ssize_t BatchRequest::encode_v2(BufferVec* bufs) const {
+int BatchRequest::encode_v2(BufferVec* bufs) const {
   const int version = 2;
 
   size_t length = 0;
@@ -77,7 +77,7 @@ ssize_t BatchRequest::encode_v2(BufferVec* bufs) const {
     const Statement* statement = *it;
 
     // <kind> [byte]
-    ssize_t buf_size = sizeof(uint8_t);
+    int buf_size = sizeof(uint8_t);
 
     // <string_or_id> [long string] | [short bytes]
     buf_size += (statement->kind() == CASS_BATCH_KIND_QUERY) ? sizeof(int32_t) : sizeof(uint16_t);
