+++ src/prepare_request.hpp
@@ -38,7 +38,7 @@ public:
   void set_query(const std::string& query) { query_ = query; }
 
 private:
-  ssize_t encode(int version, BufferVec* bufs) const;
+  int encode(int version, BufferVec* bufs) const;
 
 private:
   std::string query_;
