+++ src/startup_request.hpp
@@ -40,7 +40,7 @@ public:
   const std::string compression() const { return compression_; }
 
 private:
-  ssize_t encode(int version, BufferVec* bufs) const;
+  int encode(int version, BufferVec* bufs) const;
 
 private:
   typedef std::map<std::string, std::string> OptionsMap;
